microservices/comments-service/requirements.txt
Flask==3.0.3
google-cloud-firestore==2.16.0
google-auth==2.33.0
gunicorn==22.0.0
